// src/dto/Customer.dto.ts
import {IsEmail, IsEmpty, Length} from 'class-validator';

export class CreateCustomerInputs {
    @IsEmail()
    email: string;

    @Length(7,20)
    phone: string;

    @Length(6, 12)
    password: string;
}

export class UserLoginInputs {
    @IsEmail()
    email: string;

    @Length(6, 12)
    password: string;
}


export class EditCustomerProfileInputs {
    @Length(3, 16)
    firstName: string;

    @Length(3, 16)
    lastName: string;

    @Length(6, 16)
    address: string;
}

export interface CustomerPayload{
    _id: string;
    email: string;
    verified: boolean;
}


export class OrderInputs {
    _id: string;
    unit: number;
}